mantap sdfs

+++ app/Controllers/Buka_Order.php
@@ -903,8 +903,8 @@ class Buka_Order extends Controller
             $h_beli = $barang['harga'];
             $target_id = $this->userData['id_toko'];
 
-            $cols = 'ref,jenis,id_barang,kode_barang,id_sumber,id_target,harga_beli,qty';
-            $vals = "'" . $ref . "',0," . $id_barang . ",'" . $b_code . "','" . $id_sumber . "','" . $target_id . "'," . $h_beli . "," . $qty;
+            $cols = 'ref,jenis,id_barang,kode_barang,id_sumber,id_target,harga_beli,qty,pid';
+            $vals = "'" . $ref . "',0," . $id_barang . ",'" . $b_code . "','" . $id_sumber . "','" . $target_id . "'," . $h_beli . "," . $qty . "," . $do['id_order_data'];
             $do = $this->db(0)->insertCols('master_mutasi', $cols, $vals);
 
             if ($do['errno'] <> 0) {
@@ -950,6 +950,11 @@ class Buka_Order extends Controller
          $where = "id_order_data =" . $id_order;
       }
 
+      if ($cek_price_lock['ref'] <> "") {
+         echo "Tidak dapat dihapus, silahkan lakukan cancel";
+         exit();
+      }
+
       $do = $this->db(0)->delete_where('order_data', $where);
       if ($do['errno'] == 0) {
          $this->model('Log')->write($this->userData['user'] . " Delete Order Produksi Success!");
@@ -972,6 +977,11 @@ class Buka_Order extends Controller
          }
       } else {
          $where = "id =" . $id;
+      }
+
+      if ($cek_price_lock['ref'] <> "") {
+         echo "Tidak dapat dihapus, silahkan lakukan cancel";
+         exit();
       }
 
       $do = $this->db(0)->delete_where('master_mutasi', $where);
